Renamed a folder  that houses reusable components from 'Reusables' to 'Shared'.
