+++ MagnusApp/MagnusApp.Client/Pages/Shared/DataForm.razor
