+++ MagnusApp/MagnusApp.Client/Pages/Shared/DisplayProject.razor
