+++ MagnusApp/MagnusApp.Client/_Imports.razor
@@ -16,4 +16,4 @@
 @using Syncfusion.Blazor.Popups
 @using Syncfusion.Blazor.Notifications
 @using Syncfusion.Blazor.Spinner
-@using MagnusApp.Client.Pages.Reusables
+@using MagnusApp.Client.Pages.Shared
